proj2/requirements.txt
asgiref==3.5.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
djangorestframework==3.13.1
jwt==1.3.1
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2

celery~=5.2.6